requirements.txt
Flask==2.2.3
easyocr==1.6.2
torch==2.0.1
torchvision==0.15.2
transformers==4.27.4
Pillow==9.5.0
opencv-python==4.7.0.72
pytesseract==0.3.10